feat(#1237): widen ParameterCount int -> long across base classes + chunked api
closes #1237.

multi-day mechanical refactor that widens
`iparameterizable<t,tinput,toutput>.parametercount` and
`ilayer<t>.parametercount` from `int` to `long`, matching pytorch's
`c10::tensorimpl::numel()` int64_t convention. foundation-scale models
in the codebase (hidream full / pg3 / mj7 8b, sd3.5 large 7.6b, sora /
mochi1 5.4b, hunyuanvideo 4.5b, flux1/2/omnigen2 3.6b, sana large 3.6b,
veo 3.1b, kling 1.8b) overflow `int.maxvalue` (2.147b) in the
aggregate parametercount; this widening fixes the silent wraparound.

scope:

1. **interface widening**:
   - `iparameterizable<t,tinput,toutput>.parametercount` (`src/interfaces/iparameterizable.cs`)
   - `ilayer<t>.parametercount` (`src/interfaces/ilayer.cs`)
   - `isslmethod<t>.parametercount` (`src/selfsupervisedlearning/isslmethod.cs`)
   - `iprojectorhead<t>.parametercount` (`src/selfsupervisedlearning/iprojectorhead.cs`)

2. **base class widening**:
   - `modelbase<t,tinput,toutput>.parametercount`
   - `automlmodelbase<t,tinput,toutput>.parametercount`
   - `causalmodelbase<t,tinput,toutput>.parametercount`
   - `classifierbase<t>.parametercount`
   - `multilabelclassifierbase<t>.parametercount`
   - `clusteringbase<t,tinput,toutput>.parametercount`
   - `diffusionmodelbase<t,tinput,toutput>.parametercount`
   - `noisepredictorbase<t>.parametercount`
   - `vaemodelbase<t>.parametercount`
   - `neuralnetworkbase<t>.parametercount` + `getparametercount()`
   - `genetics/modelindividual<t>.parametercount`
   - `shardedmodelbase<t,tinput,toutput>.parametercount`
   - `sslmethodbase<t>.parametercount` + `getadditionalparametercount()`

3. **all 561 derived overrides** widened to `long` via mechanical sed
   over the if/abstract/virtual/override property declarations. ~1500
   call sites cast to `(int)` where the consuming variable / parameter
   stays int (correct for non-foundation-scale models that fit in int);
   ~80 sites widened to long where the value flows into another long
   accumulator (e.g. cross-network aggregations).

4. **chunked api overrides** added to the diffusion stack so callers
   can stream per-tensor without materialising a flat aggregate that
   overflows `vector.length`'s int contract:
   - `diffusionmodelbase.getparameterchunks()` — empty default; subclasses override
   - `latentdiffusionmodelbase.getparameterchunks()` — yields from noise predictor, vae, conditioner in turn
   - `noisepredictorbase.getparameterchunks()` — single-chunk default wrapping `getparameters()`
   - `vaemodelbase.getparameterchunks()` — same default
   - `ditnoisepredictor.getparameterchunks()` — per-layer chunks (patchembed, time-embed, label-embed, adaln, every block's norm/attention/mlp/cross-attention sub-layers, final norm + outputproj). dodges the foundation-scale buffer-allocation bug in `ditnoisepredictor.getparameters()` that this refactor surfaced.

5. **sora test un-skipped + extended** (`diffusionmodelcontracttests.cs`).
   the previous version was skipped behind a `[fact(skip = ...)]` per
   the project's "no placeholder tests" rule. now asserts:
   - paper-faithful component types (ditnoisepredictor + temporalvae)
   - `parametercount` is positive
   - chunked walk yields at least one tensor and accumulates past
     `int.maxvalue` into a `long` without overflow

build status:
- `dotnet build src/aidotnet.csproj`: clean (0 errors)
- `dotnet build tests/aidotnet.tests/aidotnettests.csproj`: clean (0 errors)
- sora test: passes in 12 s

residual scope tracked for follow-up:
- foundation-scale `getparameterchunks()` per-tensor streaming (rather
  than wrapping `getparameters()` flat aggregate) on noisepredictorbase
  and vaemodelbase. ditnoisepredictor already does per-layer; other
  predictors can follow the same pattern.

per-tensor invariant unchanged: `vector<t>.length` stays int. the
contract is "per-tensor < 2.1b; aggregate uses chunked walk" exactly
as the issue's acceptance criterion specifies.

+++ src/AiDotNet.Serving/Models/Federated/FederatedRunStatusResponse.cs
@@ -18,7 +18,7 @@ public class FederatedRunStatusResponse
     /// <summary>
     /// Gets or sets the expected parameter count.
     /// </summary>
-    public int ParameterCount { get; set; }
+    public long ParameterCount { get; set; }
 
     /// <summary>
     /// Gets or sets the number of joined clients.
