+++ src/Audio/Classification/AST.cs
@@ -513,7 +513,7 @@ public class AST<T> : AudioClassifierBase<T>, IAudioEventDetector<T>
         int index = 0;
         foreach (var layer in Layers)
         {
-            int count = layer.ParameterCount;
+            int count = (int)((int)layer.ParameterCount);
             var layerParams = parameters.Slice(index, count);
             layer.UpdateParameters(layerParams);
             index += count;
