+++ src/Audio/Classification/AudioSep.cs
@@ -292,7 +292,7 @@ public class AudioSep<T> : AudioClassifierBase<T>, IAudioEventDetector<T>
     public override void UpdateParameters(Vector<T> parameters)
     {
         if (!_useNativeMode) throw new NotSupportedException("UpdateParameters is not supported in ONNX mode.");
-        int idx = 0; foreach (var l in Layers) { int c = l.ParameterCount; l.UpdateParameters(parameters.Slice(idx, c)); idx += c; }
+        int idx = 0; foreach (var l in Layers) { int c = (int)l.ParameterCount; l.UpdateParameters(parameters.Slice(idx, c)); idx += c; }
     }
 
     protected override Tensor<T> PreprocessAudio(Tensor<T> rawAudio)
