+++ src/Audio/Classification/CLAP.cs
@@ -395,7 +395,7 @@ public class CLAP<T> : AudioClassifierBase<T>, IAudioEventDetector<T>
         int idx = 0;
         foreach (var l in Layers)
         {
-            int c = l.ParameterCount;
+            int c = (int)l.ParameterCount;
             l.UpdateParameters(parameters.Slice(idx, c));
             idx += c;
         }
