+++ src/Audio/Classification/HTSAT.cs
@@ -347,7 +347,7 @@ public class HTSAT<T> : AudioClassifierBase<T>, IAudioEventDetector<T>
         int index = 0;
         foreach (var layer in Layers)
         {
-            int count = layer.ParameterCount;
+            int count = (int)((int)layer.ParameterCount);
             layer.UpdateParameters(parameters.Slice(index, count));
             index += count;
         }
