+++ src/Diffusion/ImageEditing/InstructPix2PixModel.cs
@@ -129,7 +129,7 @@ public class InstructPix2PixModel<T> : LatentDiffusionModelBase<T>
     /// <inheritdoc />
     public override int LatentChannels => LATENT_CHANNELS;
     /// <inheritdoc />
-    public override int ParameterCount => _unet.ParameterCount + _vae.ParameterCount;
+    public override long ParameterCount => _unet.ParameterCount + _vae.ParameterCount;
 
     /// <summary>Gets the default image guidance scale for balancing image fidelity.</summary>
     public double DefaultImageGuidance => DEFAULT_IMAGE_GUIDANCE;
@@ -252,7 +252,7 @@ public class InstructPix2PixModel<T> : LatentDiffusionModelBase<T>
         {
             Name = "InstructPix2Pix", Version = "1.0",
             Description = "InstructPix2Pix instruction-based image editing with dual guidance",
-            FeatureCount = ParameterCount, Complexity = ParameterCount
+            FeatureCount = (int)ParameterCount, Complexity = (int)ParameterCount
         };
         metadata.SetProperty("architecture", "sd15-8ch-input-dual-guidance");
         metadata.SetProperty("input_channels", INPUT_CHANNELS);
