+++ src/Diffusion/NoisePredictors/FluxDoubleStreamPredictor.cs
@@ -77,7 +77,7 @@ public class FluxDoubleStreamPredictor<T> : NoisePredictorBase<T>
     /// <inheritdoc />
     public override int ContextDimension => _contextDim;
     /// <inheritdoc />
-    public override int ParameterCount { get; }
+    public override long ParameterCount { get; }
 
     /// <summary>
     /// Initializes a new FLUX double-stream predictor.
@@ -124,10 +124,10 @@ public class FluxDoubleStreamPredictor<T> : NoisePredictorBase<T>
 
     private int CalculateParameterCount()
     {
-        int count = _patchEmbed.ParameterCount;
-        foreach (var block in _doubleBlocks) count += block.ParameterCount;
-        foreach (var block in _singleBlocks) count += block.ParameterCount;
-        count += _finalLayer.ParameterCount;
+        int count = (int)((int)_patchEmbed.ParameterCount);
+        foreach (var block in _doubleBlocks) count += (int)((int)block.ParameterCount);
+        foreach (var block in _singleBlocks) count += (int)((int)block.ParameterCount);
+        count += (int)((int)_finalLayer.ParameterCount);
         return count;
     }
 
@@ -205,7 +205,7 @@ public class FluxDoubleStreamPredictor<T> : NoisePredictorBase<T>
 
     private static int SetParams(DenseLayer<T> layer, Vector<T> parameters, int offset)
     {
-        int count = layer.ParameterCount;
+        int count = (int)((int)layer.ParameterCount);
         layer.SetParameters(parameters.GetSubVector(offset, count));
         return offset + count;
     }
