+++ src/Diffusion/SuperResolution/UpscaleAVideoModel.cs
@@ -232,7 +232,7 @@ public class UpscaleAVideoModel<T> : VideoDiffusionModelBase<T>
     public override bool SupportsVideoToVideo => true;
 
     /// <inheritdoc />
-    public override int ParameterCount { get { EnsureInitialized(); return _videoUNet.GetParameters().Length + _temporalVAE.GetParameters().Length; } }
+    public override long ParameterCount { get { EnsureInitialized(); return _videoUNet.GetParameters().Length + _temporalVAE.GetParameters().Length; } }
 
     /// <summary>
     /// Gets the video upscale factor (4x).
@@ -457,8 +457,8 @@ public class UpscaleAVideoModel<T> : VideoDiffusionModelBase<T>
             Name = "Upscale-A-Video",
             Version = "1.0",
             Description = "Upscale-A-Video temporally consistent video super-resolution with diffusion",
-            FeatureCount = ParameterCount,
-            Complexity = ParameterCount
+            FeatureCount = (int)ParameterCount,
+            Complexity = (int)ParameterCount
         };
 
         metadata.SetProperty("architecture", "temporal-sr-diffusion");
